# docs/openapi/flowcept-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flowcept Webservice API",
    "description": "Read-only REST API for Flowcept provenance data. Provides workflows, tasks, and objects endpoints with query support.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/workflows": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "List Workflows",
        "description": "List workflows with optional basic filters.",
        "operationId": "list_workflows_api_v1_workflows_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User"
            }
          },
          {
            "name": "campaign_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campaign Id"
            }
          },
          {
            "name": "parent_workflow_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Workflow Id"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "filter_json",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter Json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Workflow",
        "description": "Get a workflow by id.",
        "operationId": "get_workflow_api_v1_workflows__workflow_id__get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Workflow Api V1 Workflows  Workflow Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/query": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Query Workflows",
        "description": "Run an advanced read-only workflows query.",
        "operationId": "query_workflows_api_v1_workflows_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/reports/provenance-card/download": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Download Workflow Provenance Card",
        "description": "Generate and download a workflow provenance card markdown file.",
        "operationId": "download_workflow_provenance_card_api_v1_workflows__workflow_id__reports_provenance_card_download_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AggregationSpec": {
        "properties": {
          "operator": {
            "type": "string",
            "enum": [
              "avg",
              "sum",
              "min",
              "max"
            ],
            "title": "Operator"
          },
          "field": {
            "type": "string",
            "minLength": 1,
            "title": "Field"
          }
        },
        "type": "object",
        "required": [
          "operator",
          "field"
        ],
        "title": "AggregationSpec",
        "description": "Aggregation operator and source field."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ListResponse": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "count",
          "limit"
        ],
        "title": "ListResponse",
        "description": "Generic list envelope for collection endpoints."
      },
...
      "QueryRequest": {
        "properties": {
          "filter": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filter"
          },
          "projection": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection"
          },
          "limit": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Limit",
            "default": 100
          },
          "sort": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SortSpec"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort"
          },
          "aggregation": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AggregationSpec"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aggregation"
          },
          "remove_json_unserializables": {
            "type": "boolean",
            "title": "Remove Json Unserializables",
            "default": true
          }
        },
        "type": "object",
        "title": "QueryRequest",
        "description": "Read-only query payload."
      },
      "SortSpec": {
        "properties": {
          "field": {
            "type": "string",
            "minLength": 1,
            "title": "Field"
          },
          "order": {
            "type": "integer",
            "enum": [
              1,
              -1
            ],
            "title": "Order",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "field"
        ],
        "title": "SortSpec",
        "description": "Sort field/order pair."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}